requirements.txt
azure-ai-formrecognizer==3.3.3
google-genai==0.3.0
python-dotenv==1.0.1
